+++ guillotina/tests/test_swagger.py
@@ -1,5 +1,6 @@
 from openapi_spec_validator import validate_v3_spec
 
+import json
 import os
 import pytest
 
@@ -35,7 +36,9 @@ async def test_get_swagger_index(container_requester):
 @pytest.mark.app_settings(SWAGGER_SETTINGS)
 async def test_validate_swagger_definition(container_requester):
     async with container_requester as requester:
-        for path in ("/", "/db", "/db/guillotina"):
+        await requester("POST", "/db/guillotina", data=json.dumps({"@type": "Folder", "id": "folder"}))
+        await requester("POST", "/db/guillotina", data=json.dumps({"@type": "Item", "id": "item"}))
+        for path in ("/", "/db", "/db/guillotina", "/db/guillotina/folder", "/db/guillotina/item"):
             resp, status = await requester("GET", os.path.join(path, "@swagger"))
             assert status == 200
             validate_v3_spec(resp)